IconCache: get the proper theme path based on the fact we're using a themed icon or not
If an app uses QIcon::fromTheme then we only have the icon name of it, and thus
we need to return a valid icon theme path. Now, it's not possible to return the
whole list of dirs, so we just fallback to the home icons folder, so that users
might in case add icons there.

In $SNAP world, when using desktop-launcher ~/.local/share/icons also contains
all the themes in from /usr/share/icons and potentially more (depending on the
wrapper used for launching) so, we can be quite flexible with it.